# Conflicts:
#	.gradle/2.10/taskArtifacts/cache.properties.lock
#	.gradle/2.10/taskArtifacts/fileHashes.bin
#	.gradle/2.10/taskArtifacts/fileSnapshots.bin
#	.gradle/2.10/taskArtifacts/outputFileStates.bin
#	.gradle/2.10/taskArtifacts/taskArtifacts.bin
#	.gradle/2.10/tasks/_tuyaTest_compileDebugJavaWithJavac/localClassSetAnalysis/localClassSetAnalysis.bin
#	.gradle/2.10/tasks/_tuyaTest_compileDebugJavaWithJavac/localClassSetAnalysis/localClassSetAnalysis.lock
#	.gradle/2.10/tasks/_tuyaTest_compileDebugJavaWithJavac/localJarClasspathSnapshot/localJarClasspathSnapshot.bin
#	.gradle/2.10/tasks/_tuyaTest_compileDebugJavaWithJavac/localJarClasspathSnapshot/localJarClasspathSnapshot.lock
#	.gradle/2.14.1/taskArtifacts/cache.properties
#	.gradle/2.14.1/taskArtifacts/cache.properties.lock
#	.gradle/2.14.1/taskArtifacts/fileHashes.bin
#	.gradle/2.14.1/taskArtifacts/fileSnapshots.bin
#	.gradle/2.14.1/taskArtifacts/fileSnapshotsToTreeSnapshotsIndex.bin
#	.gradle/2.14.1/taskArtifacts/taskArtifacts.bin
#	.gradle/2.14.1/tasks/_tuyaTest_compileDebugJavaWithJavac/localClassSetAnalysis/localClassSetAnalysis.bin
#	.gradle/2.14.1/tasks/_tuyaTest_compileDebugJavaWithJavac/localClassSetAnalysis/localClassSetAnalysis.lock
#	.gradle/2.14.1/tasks/_tuyaTest_compileDebugJavaWithJavac/localJarClasspathSnapshot/localJarClasspathSnapshot.bin
#	.gradle/2.14.1/tasks/_tuyaTest_compileDebugJavaWithJavac/localJarClasspathSnapshot/localJarClasspathSnapshot.lock
#	build.gradle
#	gradle/wrapper/gradle-wrapper.properties
#	tuyaTest/build.gradle
#	tuyaTest/src/main/AndroidManifest.xml
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/DeviceListAdapter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/ShareActivity.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/ShareListAdapter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/TuyaSmartApp.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/activity/BaseActivity.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/activity/DeviceCommonActivity.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/activity/DpSendActivity.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/activity/HomeActivity.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/activity/SplashActivity.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/activity/SwitchActivity.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/adapter/CommonDeviceAdapter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/adapter/GroupDeviceAdapter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/adapter/SchemaListAdapter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/env/TuyaApiUrlProvider.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/fragment/BaseFragment.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/fragment/DeviceListFragment.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/model/ISharedModel.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/model/SharedModel.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/presenter/DeviceCommonPresenter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/presenter/DeviceListFragmentPresenter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/presenter/SharedMemberAddPresenter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/presenter/SwitchPresenter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/test/presenter/DeviceTestPresenter.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/utils/CommonUtil.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/utils/ViewUtils.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/utils/WidgetUtils.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/view/ISwitchView.java
#	tuyaTest/src/main/java/com/tuya/smart/android/demo/widget/PagerSlidingTabStrip.java
#	tuyaTest/src/main/res/layout/activity_ap.xml
#	tuyaTest/src/main/res/layout/activity_common_debug.xml
#	tuyaTest/src/main/res/layout/activity_switch.xml
#	tuyaTest/src/main/res/layout/fragment_personal_center.xml
#	tuyaTest/src/main/res/layout/toolbar_top_view.xml
#	tuyaTest/src/main/res/layout/ty_progress_dialog_h.xml
#	tuyaTest/src/main/res/menu/toolbar_top_smart_device.xml
#	tuyaTest/src/main/res/values-zh/strings.xml
#	tuyaTest/src/main/res/values/dimens.xml
#	tuyaTest/src/main/res/values/ids.xml
#	tuyaTest/src/main/res/values/strings.xml
#	tuyaTest/src/main/res/values/styles.xml
#	tuyaTest/tuyaTest.iml

+++ tuyaTest/src/main/java/com/tuya/smart/android/demo/model/ISharedModel.java
@@ -14,5 +14,5 @@ public interface ISharedModel {
 
     void getReceivedList();
 
-    void updateMName(int id, String mname);
+    void updateMName(long id, String mname);
 }
